Added python worker for python script files. Instead of calling python repeatedly, import the main function of the script file and iterate over file list within python. This greatly improves numer of iterations since python only needs to initialize and import once.
Changes needed: set the option '--python_script' when calling amdtk_run.
The python script needs to have a main function that accepts one parameter in the from of sys.argv.
When using the argparser only 4 lines of code have to be changed:
 - main() --> main(script_args=None)
 - parser.parse_args() --> parser.parse_args(script_args)
 - Delete else statement in check for __name__ == '__main__'

One a ramdisk with one task the speed for amdtk_ploop_exp is roughly 1.5 to 2 times faster.
One our cluster with 500 Jobs and 37 different models the improvement is roughly a factor of 20.

This higly depends on the setup. Initializing python and importing all modules can be a huge bootleneck under high load and inefficint network file systems.
Depending on the setup python runs sevral hundred calls to the file system to find all files to be imported.