Recipes xitsonga (#13)
* Fix install instructions:
 - Added descriptions on how to install prerequisites
  - HTK
  - Anaconda
  - pywrapfst

Included patch for pywarpfst installation to work with Python3

* Added parallel environment to work with array jobs on OpenCCS

* Function for CTM file reading

* Conversion script from ctm to mlf

* Initial xitsonga recipe

* Repair check for symbolic link: Check first for link then for file type
Added ability to split wav files by using sox analog to sph2pipe
Included instalation instructions for sox

* Added python worker for python script files.
Instead of calling python repeatedly, import the main function of the script file and iterate over file list within python.
This greatly improves numer of iterations since python only needs to initialize and import once.

Changes needed: set the option '--python_script' when calling amdtk_run.
The python script needs to have a main function that accepts one parameter in the from of sys.argv.
When using the argparser only 4 lines of code have to be changed:
 - main() --> main(script_args=None)
 - parser.parse_args() --> parser.parse_args(script_args)
 - Delete else statement in check for __name__ == '__main__'

One a ramdisk with one task the speed for amdtk_ploop_exp is roughly 1.5 to 2 times faster.
One our cluster with 500 Jobs and 37 different models the improvement is roughly a factor of 20.

This higly depends on the setup. Initializing python and importing all modules can be a huge bootleneck under high load and inefficint network file systems.
Depending on the setup python runs sevral hundred calls to the file system to find all files to be imported.

* Rearange sentences

* Add parameter description

* minor fixup - removing unneeded lines

* add cython - seems to be needed to run hmm

* Fixup long lines

* Fix install instructions:
 - Added descriptions on how to install prerequisites
  - HTK
  - Anaconda
  - pywrapfst

Included patch for pywarpfst installation to work with Python3

* Rearange sentences

* minor fixup - removing unneeded lines

* add cython - seems to be needed to run hmm

* Added parallel environment to work with array jobs on OpenCCS

* Function for CTM file reading

* Conversion script from ctm to mlf

* Initial xitsonga recipe

* Repair check for symbolic link: Check first for link then for file type
Added ability to split wav files by using sox analog to sph2pipe
Included instalation instructions for sox

* Added python worker for python script files.
Instead of calling python repeatedly, import the main function of the script file and iterate over file list within python.
This greatly improves numer of iterations since python only needs to initialize and import once.

Changes needed: set the option '--python_script' when calling amdtk_run.
The python script needs to have a main function that accepts one parameter in the from of sys.argv.
When using the argparser only 4 lines of code have to be changed:
 - main() --> main(script_args=None)
 - parser.parse_args() --> parser.parse_args(script_args)
 - Delete else statement in check for __name__ == '__main__'

One a ramdisk with one task the speed for amdtk_ploop_exp is roughly 1.5 to 2 times faster.
One our cluster with 500 Jobs and 37 different models the improvement is roughly a factor of 20.

This higly depends on the setup. Initializing python and importing all modules can be a huge bootleneck under high load and inefficint network file systems.
Depending on the setup python runs sevral hundred calls to the file system to find all files to be imported.

* Add parameter description

* Fixup long lines

* Make pywrapfst optional. It is not actively used in the current implementation and recipies. It is only present in the functions:
 - toFst (hmm graph class)
 - samplePathFromFst (lm controller module)
 - readHtkLattice (internal io module)

* Repair check for symbolic link: Check first for link then for file type
Added ability to split wav files by using sox analog to sph2pipe
Included instalation instructions for sox

* Make sox optional since it is not yet used in any active examples.

* Enable phone loop labeling to use the python file list worker

* reintroduce --keyfile, --directory and --fname_format option

* Add option to ignore 'sil' label in proposed mlf
Add option to correct for offset in segmented files (e.g. files with silence cut off)

* Adjust example scripts including bigram retraining and evaluation.
Work with segmented files (cutting of silence) in the example.
We will need a (better) silence detector in case we want to work with data without silence annotation.
Since the silences in the data are very obvious, we assume a good silence detector will not deliver a much different result.

* make phone loop post work with python file list worker

* lattice generation and scoring from model trained with bigram lm

* unigram lattice generation and scoring